Add Processing and NeedsIndexing checks to search
CloudsearchProcessingException and CloudsearchNeedsIndexingException
are now emitted in the constructor of a SearchConnection as long as
loose=True and needs_integrity=False. These are both defaulted to not
break backwards compatibility, but Domain.get_search_service() was
modified to take these and pass them along if you'd like some extra
sanity checking.